requirements.txt
requests==2.31.0
PyTelegramBotAPI==4.14.0
fake-useragent==1.4.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.0
playwright==1.53.0